Add a backup test in iso file
This uses the ISO output. A ramdisk is used to emulate the CD-ROM drive.
The ISO image with backup gets written to the ramdisk after ReaR
recovery starts and before it starts reformatting the disk (where the
ISO image is stored) to ensure that the backup persists until we need to
restore it. The changes relative to make-backup-and-restore-bios are
described in "A. Voznia: Automation of testing of operating system
backup and recovery (2023)" https://dspace.cvut.cz/handle/10467/107238.

Co-authored-by: Pavel Cahyna <pcahyna@redhat.com>