Allow using the output path before ReaR runs
/var/lib/rear/output gets created during the ReaR run, but we will need
to use it before in order to discover the filesystem. Split
REAR_ISO_OUTPUT into two components : REAR_ISO_FHSDIR, which always
exists on the system, and REAR_ISO_SUBDIR, which gets created by
installing or running ReaR. The former will be used to discover the
filesystem.

This assumes that the ReaR directory will not be on a different
filesystem than the standard directory that contains it, which is
usually a safe assumption.