Adapt to /var possibly being a separate filesystem
Do not assume that ReaR output is on the root filesystem: change
ROOT_DISK to OUTPUT_DISK etc. Moreover, the output may be on a subvolume
(btrfs), so adapt to that.

This fixes booting of the rescue ISO on Fedora 40 and later distributed
cloud images, as the cloud images now changed their layout a bit and
have separate /var as a btrfs subvolume due to the switch to Kiwi:
https://fedoraproject.org/wiki/Changes/KiwiBuiltCloudImages